+++ app/Console/Kernel.php
@@ -56,6 +56,8 @@ class Kernel extends ConsoleKernel
                 }
             }
         }
+
+        $schedule->command('telescope:prune')->daily();
     }
 
     /**