+++ app/Exceptions/Handler.php
@@ -5,6 +5,7 @@ namespace App\Exceptions;
 use BristolSU\Support\ActivityInstance\Contracts\ActivityInstanceResolver;
 use BristolSU\Support\ActivityInstance\Contracts\DefaultActivityInstanceGenerator;
 use BristolSU\Support\ActivityInstance\Exceptions\NotInActivityInstanceException;
+use BristolSU\Support\Authentication\Contracts\Authentication;
 use BristolSU\Support\Authentication\Contracts\ResourceIdGenerator;
 use BristolSU\Support\Authorization\Exception\ActivityDisabled;
 use BristolSU\Support\Authorization\Exception\ActivityRequiresAdmin;
@@ -16,14 +17,12 @@ use BristolSU\Support\Authorization\Exception\IncorrectLogin;
 use BristolSU\Support\Authorization\Exception\ModuleInactive;
 use BristolSU\Support\Authorization\Exception\ModuleInstanceDisabled;
 use Exception;
-use GuzzleHttp\Exception\ClientException;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Support\Str;
-use Symfony\Component\HttpKernel\Exception\HttpException;
+use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
 use Throwable;
 
 class Handler extends ExceptionHandler
@@ -36,8 +35,9 @@ class Handler extends ExceptionHandler
     protected $dontReport = [
         ActivityRequiresParticipant::class,
         ActivityRequiresAdmin::class,
-        ModuleInactive::class,
         NotInActivityInstanceException::class,
+
+        ModuleInactive::class,
         ActivityDisabled::class,
         ModuleInstanceDisabled::class
     ];
@@ -53,88 +53,101 @@ class Handler extends ExceptionHandler
     ];
 
     /**
-     * Report or log an exception.
+     * Register the exception handling callbacks for the application.
      *
-     * @param Throwable $exception
      * @return void
-     * @throws Throwable
      */
-    public function report(Throwable $exception)
+    public function register()
     {
-        parent::report($exception);
+        $this->renderable(function (ActivityRequiresParticipant $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => 'You do not have access in your current role.'], $exception->getCode());
+            }
+            return redirect()->route('login.participant', [
+                'activity_slug' => $exception->getActivity()->slug,
+                'redirect' => $request->fullUrl()
+            ]);
+        });
+        $this->renderable(function (ActivityRequiresAdmin $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => 'You do not have access to the admin side in your current role.'], $exception->getCode());
+            }
+            return redirect()->route('login.admin', [
+                'activity_slug' => $exception->getActivity()->slug,
+                'redirect' => $request->fullUrl()
+            ]);
+        });
+        $this->renderable(function (ModuleInactive $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => 'This module is currently locked for you.'], $exception->getCode());
+            }
+        });
+        $this->renderable(function (NotInActivityInstanceException $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => 'You must be in an activity instance to access this page.'], $exception->getCode());
+            }
+            $activity = $request->route('activity_slug');
+            try {
+                $activityInstance = app(DefaultActivityInstanceGenerator::class)
+                    ->generate(
+                        $activity,
+                        $activity->activity_for,
+                        app(ResourceIdGenerator::class)->fromString($activity->activity_for)
+                    );
+            } catch (Exception $e) {
+                if($activity->activity_for === 'user') {
+                    throw new ActivityRequiresUser('', 0, null, $activity);
+                }
+                if($activity->activity_for === 'group') {
+                    throw new ActivityRequiresGroup('', 0, null, $activity);
+                }
+                if($activity->activity_for === 'role') {
+                    throw new ActivityRequiresRole('', 0, null, $activity);
+                }
+            }
+            app(ActivityInstanceResolver::class)->setActivityInstance($activityInstance);
+            return redirect()->to($request->fullUrl());
+        });
+        $this->renderable(function (ActivityDisabled $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => sprintf('%s has been disabled', $exception->activity()->name)], $exception->getCode());
+            }
+        });
+        $this->renderable(function (ModuleInstanceDisabled $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => 'This module has been disabled.'], $exception->getCode());
+            }
+        });
+        $this->renderable(function (IncorrectLogin $exception, Request $request) {
+            if($request->expectsJson()) {
+                return response()->json(['message' => 'You do not have access to the credentials you are using.'], $exception->getCode());
+            }
+        });
+    }
+
+    protected function getHttpExceptionView(HttpExceptionInterface $e)
+    {
+        return 'errors.error';
     }
 
     /**
-     * Render an exception into an HTTP response.
+     * Get the default context variables for logging.
      *
-     * - Redirect to participant login page if $e instance of ActivityRequiresParticipant
-     * - Redirect to admin login page if $e instance of ActivityRequiresAdmin
-     * - Return a 403 HttpException if $e instanceof ModuleInactive
-     * - Log into the default activity instance if possible if $e instanceof NotInActivityInstanceException
-     *
-     * @param Request $request
-     * @param Throwable $exception
-     * @return Response|RedirectResponse|JsonResponse
-     * @throws ActivityRequiresGroup
-     * @throws ActivityRequiresRole
-     * @throws ActivityRequiresUser
+     * @return array
      */
-    public function render($request, Throwable $exception)
+    protected function context()
     {
-        if (!$request->expectsJson()) {
-
-            if ($exception instanceof ActivityRequiresParticipant) {
-                return redirect()->route('login.participant', [
-                    'activity_slug' => $exception->getActivity()->slug,
-                    'redirect' => $request->fullUrl()
-                ]);
-            }
-            if($exception instanceof ActivityRequiresAdmin) {
-                return redirect()->route('login.admin', [
-                    'activity_slug' => $exception->getActivity()->slug,
-                    'redirect' => $request->fullUrl()
-                ]);
-            }
-
-            if($exception instanceof ModuleInactive) {
-                abort(403);
-            }
-            if($exception instanceof NotInActivityInstanceException) {
-                $activity = $request->route('activity_slug');
-                try {
-                    $activityInstance = app(DefaultActivityInstanceGenerator::class)
-                        ->generate(
-                            $activity,
-                            $activity->activity_for,
-                            app(ResourceIdGenerator::class)->fromString($activity->activity_for)
-                        );
-                } catch (Exception $e) {
-                    if($activity->activity_for === 'user') {
-                        throw new ActivityRequiresUser('', 0, null, $activity);
-                    }
-                    if($activity->activity_for === 'group') {
-                        throw new ActivityRequiresGroup('', 0, null, $activity);
-                    }
-                    if($activity->activity_for === 'role') {
-                        throw new ActivityRequiresRole('', 0, null, $activity);
-                    }
-                }
-                app(ActivityInstanceResolver::class)->setActivityInstance($activityInstance);
-                return redirect()->to($request->fullUrl());
-            }
-            if($exception instanceof ActivityDisabled) {
-                return response()->view('errors.activity_disabled', ['activity' => $exception->activity()]);
-            } if($exception instanceof ModuleInstanceDisabled) {
-                return response()->view('errors.module_instance_disabled', ['moduleInstance' => $exception->moduleInstance()]);
-            }
-        } else {
-            if($exception instanceof IncorrectLogin) {
-                return response()->json(['message' => $exception->getMessage()], $exception->getStatusCode());
-            }
+        $context = [];
+        $auth = app(Authentication::class);
+        if($auth->hasUser()) {
+            $context['control_user_id'] = $auth->getUser()->id();
         }
-
-        // TODO Handle exceptions above if expects json
-
-        return parent::render($request, $exception);
+        if($auth->hasGroup()) {
+            $context['control_group_id'] = $auth->getGroup()->id();
+        }
+        if($auth->hasRole()) {
+            $context['control_role_id'] = $auth->getRole()->id();
+        }
+        return array_merge(parent::context(), $context);
     }
 }