+++ app/Http/Controllers/Api/ModuleInstanceController.php
@@ -30,4 +30,16 @@ class ModuleInstanceController extends Controller
         ]));
     }
 
+    /**
+     * @param ModuleInstance $moduleInstance
+     * @return \Illuminate\Http\JsonResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
+    public function destroy(ModuleInstance $moduleInstance)
+    {
+        $this->authorize('view-management');
+        $moduleInstance->delete();
+        return response()->json($moduleInstance);
+    }
+
 }