+++ resources/views/management/activity/show.blade.php
@@ -1,12 +1,16 @@
-@extends('layouts.settings')
+@extends('layouts.legacy')
 
-@section('settings-title', $activity->name)
+@section('title', $activity->name)
 
-@section('settings-content')
-
+@section('app-content')
     <activity-show
+        @can('view-management')
+            :can-delete=true
+        @endcan
         :activity="{{$activity}}">
 
     </activity-show>
 
 @endsection
+
+