+++ resources/views/management/logic/show.blade.php
@@ -1,8 +1,8 @@
-@extends('layouts.settings')
+@extends('layouts.legacy')
 
-@section('settings-title', $logic->name)
+@section('title', $logic->name)
 
-@section('settings-content')
+@section('app-content')
 
     <logic-show
         :logic="{{$logic}}">
@@ -10,3 +10,5 @@
     </logic-show>
 
 @endsection
+
+