+++ resources/views/portal/control.blade.php
@@ -10,3 +10,7 @@
 @push('scripts')
     <script src="{{mix('js/control.js')}}"></script>
 @endpush
+
+@push('styles')
+    <link href="{{ mix('css/app.css') }}" rel="stylesheet">
+@endpush